refactor: move CommandUtils to handlers
