+++ src/main/java/com/festp/handlers/WhisperHandler.java
@@ -18,7 +18,6 @@ import org.bukkit.event.server.ServerCommandEvent;
 import com.festp.Chatter;
 import com.festp.Logger;
 import com.festp.config.Config;
-import com.festp.parsing.CommandUtils;
 import com.festp.parsing.StyledMessage;
 import com.festp.parsing.StyledMessageParser;
 import com.festp.parsing.TextStyle;
